+++ src/entrypoints/app.ts
@@ -1,4 +1,4 @@
 import "@webcomponents/scoped-custom-element-registry/scoped-custom-element-registry.min";
 import "../layouts/home-assistant";
 
-import("../resources/ha-style");
+import("../resources/append-ha-style");
