+++ src/resources/append-ha-style.ts
@@ -1,0 +1,5 @@
+import { themeStyles } from "./theme/theme";
+
+const styleElement = document.createElement("style");
+styleElement.textContent = themeStyles;
+document.head.append(styleElement);
