+++ src/resources/styles-data.ts
@@ -55,187 +55,3 @@ export const darkStyles = {
     "invert(.9) hue-rotate(170deg) brightness(1.5) contrast(1.2) saturate(.3)",
   "disabled-color": "#464646",
 } as const;
-
-export const derivedStyles = {
-  /* Vaadin typography */
-  "material-h6-font-size": "var(--ha-font-size-m)",
-  "material-small-font-size": "var(--ha-font-size-xs)",
-  "material-caption-font-size": "var(--ha-font-size-2xs)",
-  "material-button-font-size": "var(--ha-font-size-xs)",
-
-  /* for label-badge */
-  "label-badge-red": "var(--error-color)",
-  "label-badge-blue": "var(--info-color)",
-  "label-badge-green": "var(--success-color)",
-  "label-badge-yellow": "var(--warning-color)",
-
-  /* state color */
-  "state-active-color": "var(--amber-color)",
-  "state-inactive-color": "var(--grey-color)",
-  "state-unavailable-color":
-    "var(--state-icon-unavailable-color, var(--disabled-text-color))",
-
-  /* state domain colors */
-  "state-alarm_control_panel-armed_away-color": "var(--green-color)",
-  "state-alarm_control_panel-armed_custom_bypass-color": "var(--green-color)",
-  "state-alarm_control_panel-armed_home-color": "var(--green-color)",
-  "state-alarm_control_panel-armed_night-color": "var(--green-color)",
-  "state-alarm_control_panel-armed_vacation-color": "var(--green-color)",
-  "state-alarm_control_panel-arming-color": "var(--orange-color)",
-  "state-alarm_control_panel-disarming-color": "var(--orange-color)",
-  "state-alarm_control_panel-pending-color": "var(--orange-color)",
-  "state-alarm_control_panel-triggered-color": "var(--red-color)",
-  "state-alert-off-color": "var(--orange-color)",
-  "state-alert-on-color": "var(--red-color)",
-  "state-binary_sensor-active-color": "var(--amber-color)",
-  "state-binary_sensor-battery-on-color": "var(--red-color)",
-  "state-binary_sensor-carbon_monoxide-on-color": "var(--red-color)",
-  "state-binary_sensor-gas-on-color": "var(--red-color)",
-  "state-binary_sensor-heat-on-color": "var(--red-color)",
-  "state-binary_sensor-lock-on-color": "var(--red-color)",
-  "state-binary_sensor-moisture-on-color": "var(--red-color)",
-  "state-binary_sensor-problem-on-color": "var(--red-color)",
-  "state-binary_sensor-safety-on-color": "var(--red-color)",
-  "state-binary_sensor-smoke-on-color": "var(--red-color)",
-  "state-binary_sensor-sound-on-color": "var(--red-color)",
-  "state-binary_sensor-tamper-on-color": "var(--red-color)",
-  "state-climate-auto-color": "var(--green-color)",
-  "state-climate-cool-color": "var(--blue-color)",
-  "state-climate-dry-color": "var(--orange-color)",
-  "state-climate-fan_only-color": "var(--cyan-color)",
-  "state-climate-heat-color": "var(--deep-orange-color)",
-  "state-climate-heat-cool-color": "var(--amber-color)",
-  "state-cover-active-color": "var(--purple-color)",
-  "state-device_tracker-active-color": "var(--blue-color)",
-  "state-device_tracker-home-color": "var(--green-color)",
-  "state-fan-active-color": "var(--cyan-color)",
-  "state-humidifier-on-color": "var(--blue-color)",
-  "state-lawn_mower-error-color": "var(--red-color)",
-  "state-lawn_mower-mowing-color": "var(--teal-color)",
-  "state-light-active-color": "var(--amber-color)",
-  "state-lock-jammed-color": "var(--red-color)",
-  "state-lock-locked-color": "var(--green-color)",
-  "state-lock-locking-color": "var(--orange-color)",
-  "state-lock-unlocked-color": "var(--red-color)",
-  "state-lock-unlocking-color": "var(--orange-color)",
-  "state-lock-open-color": "var(--red-color)",
-  "state-lock-opening-color": "var(--orange-color)",
-  "state-media_player-active-color": "var(--light-blue-color)",
-  "state-person-active-color": "var(--blue-color)",
-  "state-person-home-color": "var(--green-color)",
-  "state-plant-active-color": "var(--red-color)",
-  "state-siren-active-color": "var(--red-color)",
-  "state-sun-above_horizon-color": "var(--amber-color)",
-  "state-sun-below_horizon-color": "var(--indigo-color)",
-  "state-switch-active-color": "var(--amber-color)",
-  "state-update-active-color": "var(--orange-color)",
-  "state-vacuum-active-color": "var(--teal-color)",
-  "state-valve-active-color": "var(--blue-color)",
-  "state-sensor-battery-high-color": "var(--green-color)",
-  "state-sensor-battery-low-color": "var(--red-color)",
-  "state-sensor-battery-medium-color": "var(--orange-color)",
-  "state-water_heater-eco-color": "var(--green-color)",
-  "state-water_heater-electric-color": "var(--orange-color)",
-  "state-water_heater-gas-color": "var(--orange-color)",
-  "state-water_heater-heat_pump-color": "var(--orange-color)",
-  "state-water_heater-high_demand-color": "var(--deep-orange-color)",
-  "state-water_heater-performance-color": "var(--deep-orange-color)",
-
-  /* history colors */
-  "history-unknown-color": "var(--dark-grey-color)",
-
-  "state-icon-error-color": "var(--error-state-color, var(--error-color))",
-  "sidebar-text-color": "var(--primary-text-color)",
-  "sidebar-background-color": "var(--card-background-color)",
-  "sidebar-selected-text-color": "var(--primary-color)",
-  "sidebar-selected-icon-color": "var(--primary-color)",
-  "sidebar-icon-color": "rgba(var(--rgb-primary-text-color), 0.6)",
-  "switch-checked-color": "var(--primary-color)",
-  "switch-checked-button-color":
-    "var(--switch-checked-color, var(--primary-background-color))",
-  "switch-checked-track-color": "var(--switch-checked-color, #000000)",
-  "switch-unchecked-button-color":
-    "var(--switch-unchecked-color, var(--primary-background-color))",
-  "switch-unchecked-track-color": "var(--switch-unchecked-color, #000000)",
-  "slider-color": "var(--primary-color)",
-  "slider-secondary-color": "var(--light-primary-color)",
-  "slider-track-color": "var(--scrollbar-thumb-color)",
-  "label-badge-background-color": "var(--card-background-color)",
-  "label-badge-text-color": "rgba(var(--rgb-primary-text-color), 0.8)",
-  "paper-listbox-background-color": "var(--card-background-color)",
-  "paper-item-icon-active-color": "var(--state-icon-active-color)",
-  "table-header-background-color": "var(--input-fill-color)",
-  "table-row-background-color": "var(--primary-background-color)",
-  "table-row-alternative-background-color": "var(--secondary-background-color)",
-  "data-table-background-color": "var(--card-background-color)",
-  "markdown-code-background-color": "var(--primary-background-color)",
-
-  // https://github.com/material-components/material-web/blob/master/docs/theming.md
-  "mdc-theme-primary": "var(--primary-color)",
-  "mdc-theme-secondary": "var(--accent-color)",
-  "mdc-theme-background": "var(--primary-background-color)",
-  "mdc-theme-surface": "var(--card-background-color)",
-  "mdc-theme-on-primary": "var(--text-primary-color)",
-  "mdc-theme-on-secondary": "var(--text-primary-color)",
-  "mdc-theme-on-surface": "var(--primary-text-color)",
-  "mdc-theme-text-disabled-on-light": "var(--disabled-text-color)",
-  "mdc-theme-text-primary-on-background": "var(--primary-text-color)",
-  "mdc-theme-text-secondary-on-background": "var(--secondary-text-color)",
-  "mdc-theme-text-hint-on-background": "var(--secondary-text-color)",
-  "mdc-theme-text-icon-on-background": "var(--secondary-text-color)",
-  "mdc-theme-error": "var(--error-color)",
-  "app-header-text-color": "var(--text-primary-color)",
-  "app-header-background-color": "var(--primary-color)",
-  "app-theme-color": "var(--app-header-background-color)",
-  "mdc-checkbox-unchecked-color": "rgba(var(--rgb-primary-text-color), 0.54)",
-  "mdc-checkbox-disabled-color": "var(--disabled-text-color)",
-  "mdc-radio-unchecked-color": "rgba(var(--rgb-primary-text-color), 0.54)",
-  "mdc-radio-disabled-color": "var(--disabled-text-color)",
-  "mdc-tab-text-label-color-default": "var(--primary-text-color)",
-  "mdc-button-disabled-ink-color": "var(--disabled-text-color)",
-  "mdc-button-outline-color": "var(--outline-color)",
-  "mdc-dialog-scroll-divider-color": "var(--divider-color)",
-  "mdc-dialog-heading-ink-color": "var(--primary-text-color)",
-  "mdc-dialog-content-ink-color": "var(--primary-text-color)",
-
-  "mdc-text-field-idle-line-color": "var(--input-idle-line-color)",
-  "mdc-text-field-hover-line-color": "var(--input-hover-line-color)",
-  "mdc-text-field-disabled-line-color": "var(--input-disabled-line-color)",
-  "mdc-text-field-outlined-idle-border-color":
-    "var(--input-outlined-idle-border-color)",
-  "mdc-text-field-outlined-hover-border-color":
-    "var(--input-outlined-hover-border-color)",
-  "mdc-text-field-outlined-disabled-border-color":
-    "var(--input-outlined-disabled-border-color)",
-  "mdc-text-field-fill-color": "var(--input-fill-color)",
-  "mdc-text-field-disabled-fill-color": "var(--input-disabled-fill-color)",
-  "mdc-text-field-ink-color": "var(--input-ink-color)",
-  "mdc-text-field-label-ink-color": "var(--input-label-ink-color)",
-  "mdc-text-field-disabled-ink-color": "var(--input-disabled-ink-color)",
-
-  "mdc-select-idle-line-color": "var(--input-idle-line-color)",
-  "mdc-select-hover-line-color": "var(--input-hover-line-color)",
-  "mdc-select-outlined-idle-border-color":
-    "var(--input-outlined-idle-border-color)",
-  "mdc-select-outlined-hover-border-color":
-    "var(--input-outlined-hover-border-color)",
-  "mdc-select-outlined-disabled-border-color":
-    "var(--input-outlined-disabled-border-color)",
-  "mdc-select-fill-color": "var(--input-fill-color)",
-  "mdc-select-disabled-fill-color": "var(--input-disabled-fill-color)",
-  "mdc-select-ink-color": "var(--input-ink-color)",
-  "mdc-select-label-ink-color": "var(--input-label-ink-color)",
-  "mdc-select-disabled-ink-color": "var(--input-disabled-ink-color)",
-  "mdc-select-dropdown-icon-color": "var(--input-dropdown-icon-color)",
-  "mdc-select-disabled-dropdown-icon-color": "var(--input-disabled-ink-color)",
-  "ha-assist-chip-filled-container-color":
-    "rgba(var(--rgb-primary-text-color),0.15)",
-  "ha-assist-chip-active-container-color":
-    "rgba(var(--rgb-primary-color),0.15)",
-  "chip-background-color": "rgba(var(--rgb-primary-text-color), 0.15)",
-  // Vaadin
-  "material-body-text-color": "var(--primary-text-color)",
-  "material-background-color": "var(--card-background-color)",
-  "material-secondary-background-color": "var(--secondary-background-color)",
-  "material-secondary-text-color": "var(--secondary-text-color)",
-} as const;
