+++ src/resources/theme/typography.ts
@@ -1,11 +1,11 @@
-import { splitCssVars as css } from "./css-vars";
+import { extractDerivedVars } from "../../common/style/derived-css-vars";
 
-const result = css`
+const result = extractDerivedVars`
   html {
     --ha-font-family-body: Roboto, Noto, sans-serif;
     --ha-font-family-code: monospace;
     --ha-font-family-longform: ui-sans-serif, system-ui, sans-serif;
-
+    
     font-size: 14px;
     --ha-font-size-scale: 1;
     --ha-font-size-xs: calc(10px * var(--ha-font-size-scale));
